Configured the navigator for proper flow
