+++ src/App.js
@@ -2,30 +2,16 @@ import 'react-native-gesture-handler';
 import React from 'react';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
 import { ThemeProvider } from 'styled-components';
-import { Home, CurrencyList, Options, Themes } from './screens';
+import { RootNavigator } from './navigators';
 import { theme } from './theme';
-
-const Stack = createStackNavigator();
 
 const App = () => {
   return (
     <NavigationContainer>
       <SafeAreaProvider>
         <ThemeProvider theme={theme}>
-          <Stack.Navigator initialRouteName='Home'>
-            <Stack.Screen
-              name='Home'
-              component={Home}
-              options={{
-                headerShown: false,
-              }}
-            />
-            <Stack.Screen name='CurrencyList' component={CurrencyList} />
-            <Stack.Screen name='Options' component={Options} />
-            <Stack.Screen name='Themes' component={Themes} />
-          </Stack.Navigator>
+          <RootNavigator />
         </ThemeProvider>
       </SafeAreaProvider>
     </NavigationContainer>
